Adjustments in Card View
Fixed control flow in CardView.
Added padding to the shapes.
Flipped the cards horizontally to mimic real-life Set play.

Think about:
In real-life Set all shapes are the same size despite different counts. This will need to be scaled as more cards are added, but how can we makes all shapes the same size for a give number of cards?

When the cards were flipped, its possible that the Grid spacing needs to be re-optomized.